Use parent planstate as span end for top level spans (#5)
Top spans end are currently ended at the end of the parent's ExecutorRun
while nodes will rely on query instrumentation. This can lead to a
situation where a node with nested queries like ProjectSet is shorter
than its children.

We already use traced_planstate to correctly assign parent for a nested
top span. We will keep track of the traced_planstate in the span and use
it to when ending the top span.