Replace latest_trace_id by tx_start_traceparent
If the start of a transaction has a trace_context, we want to reuse it
for the whole transaction. Replace latest_trace_id by
tx_start_traceparent which will only be set at the start of a new
transaction and update it with both randomly generated trace_id and
extracted_trace_id from SQLComments.

If a statement has its own trace_context, we give it priority.

This fixes the issue #13 where latest_trace_id was not updated when
trace_id was provided by SQLComment, leading to creating spans with a 0
trace_id.