Remove query_id from tracecontext
We can grab query_id from Query or PlannedStmt in most cases. The only
place where we need to provide a queryId is for the commit span. Since
it only happens at level 0, we just need to save query_id during
ExecutorFinish to be reused in pg_tracing_xact_callback.