Allow changing otel endpoint setting without restart
Also the naptime and connection timeout settings. If no endpoint was
set at server start, though, the background worker is not started, and
changing the endpoint will have no effect. Similarly, if you set
endpoint to empty while the server is running, the background worker
will keep running, but won't do anything. Changing the registration of
the background worker in response to config changes is difficult. But
at least you can change the endpoint without a restart.